Refactor EllipticalGradient.kt out into separate file
